Add fullscreen burger menu to home page
Introduces a fullscreen navigation menu activated by a burger icon on the home page. Updates CSS and JavaScript in index.html.twig to support the new menu, and adds the menu structure to the template. Also updates .gitignore to exclude a secrets file and adds a .DS_Store file to config.